Create a menubar package and move related classes into it
